backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
flask-socketio==5.12.0
eventlet==0.33.4
openai==1.0.0
httpx==0.24.1
python-dotenv==1.0.0
requests==2.31.0
faster-whisper

